autoselect `int32`/`float64` only

+++ tests/testthat/test-info.r
@@ -41,8 +41,8 @@ test_that("h5_typeof and h5_class", {
   expect_equal(h5_typeof(file, 'ds_opaque'),   'opaque')
   expect_equal(h5_typeof(file, 'ds_enum'),     'enum')
 
-  expect_equal(h5_class(file, 'ds_uint8'),    'numeric')
-  expect_equal(h5_class(file, 'ds_int8'),     'numeric')
+  expect_equal(h5_class(file, 'ds_uint8'),    'integer')
+  expect_equal(h5_class(file, 'ds_int8'),     'integer')
   expect_equal(h5_class(file, 'ds_float16'),  'numeric')
   expect_equal(h5_class(file, 'ds_null'),     'NULL')
   expect_equal(h5_class(file, 'ds_string'),   'character')
@@ -58,7 +58,7 @@ test_that("h5_typeof and h5_class", {
   expect_equal(h5_typeof(file, '/', 'attr_uint8'), 'uint8')
   expect_equal(h5_typeof(file, '/', 'attr_null'),  'null')
 
-  expect_equal(h5_class(file, '/', 'attr_uint8'), 'numeric')
+  expect_equal(h5_class(file, '/', 'attr_uint8'), 'integer')
   expect_equal(h5_class(file, '/', 'attr_null'),  'NULL')
